Better functionality for Friendly NPC dialogue.
Added better functionality for the dialogue. Can add as many lines as you want to the list and click the button to go through it. After finishing the UI will no longer be active and the dialogue restart upon reentering the interaction zone. Made it a prefab and can add to other rooms. Make sure to add an event system in that scene when adding to other rooms.