End Dialogue and Environment FIx in Morning Room
Added functionality to change the friendly NPC to have an end dialogue once you complete the rooms and go back to the lobby room. Added an invisible wall as a barrier for the gates to the mansion in the morning scene so you can't actually go to that area and fall through the model.