Changed NPC (Still Kinda Crashes, Read Desc)
Changed the NPC initially, it would crash on the first startup from the automatic build, but every time after it would load and work perfectly fine. I then added duplicates of the NPC for the Morning Room in the copy areas for the "Portal" behavior. After this, it would crash every single time after building like with the old NPC. I removed the duplicates and it still kept crashing. I then uninstalled the app, and rechecked/reviewed player settings and now it works with no crashing after a few tests after building. Not sure what is causing the crashing the specifically. This makes me feel like it's not the NPC but something else.